+++ src/rules/no-test-imports-in-prod/__test__/project/testing/e.ts
@@ -1,0 +1,1 @@
+export const eHelper = 10;