+++ src/rules/no-test-only-imports/__test__/project/tests/j-test.ts
@@ -1,0 +1,3 @@
+import { a6 } from '../a';
+
+console.log(a6);